[REFACTOR] : 함수 순서 바꾸기 / #1
