+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:name=".features.account.SignUpActivity"
             android:exported="false" />
         <activity
-            android:name=".features.MainActivity"
+            android:name=".features.main.MainActivity"
             android:exported="false" />
     </application>
