+++ app/src/main/java/org/sopt/dosopttemplate/features/account/SignInActivity.kt
@@ -3,7 +3,6 @@ package org.sopt.dosopttemplate.features.account
 import android.app.Activity
 import android.content.Intent
 import android.os.Build
-import android.os.Bundle
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import com.example.core_ui.base.BindingActivity
@@ -11,61 +10,18 @@ import com.example.core_ui.context.snackBar
 import com.example.core_ui.context.toast
 import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.databinding.ActivitySignInBinding
-import org.sopt.dosopttemplate.features.MainActivity
+import org.sopt.dosopttemplate.features.account.model.User
+import org.sopt.dosopttemplate.features.main.MainActivity
 import org.sopt.dosopttemplate.features.util.Account.SIGN_UP_INFORMATION
 
 class SignInActivity : BindingActivity<ActivitySignInBinding>(R.layout.activity_sign_in) {
     private lateinit var resultLauncher: ActivityResultLauncher<Intent>
     private var savedSignUpInformation: User? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setResultSignUpInformation()
-    }
-
     override fun initView() {
         setResultSignUpInformation()
         setClickEventOnSignUpLabelButton()
         setClickEventOnSignInLabelButton()
-    }
-
-    private fun setClickEventOnSignUpLabelButton() {
-        binding.btnSignInSignupLabel.setOnClickListener {
-            navigateToSignUp()
-        }
-    }
-
-    private fun setClickEventOnSignInLabelButton() {
-        binding.btnSignInSigninLabel.setOnClickListener {
-            savedSignUpInformation?.let {
-                if (isSignInSuccessful(it)) handleSignInSuccess() else showSignInErrorMessage()
-            }
-        }
-    }
-
-    private fun isSignInSuccessful(user: User): Boolean {
-        return checkIdIdentification(user.id) && checkPwIdentification(user.pw)
-    }
-
-    private fun showSignInErrorMessage() {
-        snackBar(binding.root) { getString(R.string.error_message_invalid_sign_in) }
-    }
-
-    private fun handleSignInSuccess() {
-        toast(getString(R.string.error_message_valid_sign_in))
-        navigateToMain()
-    }
-
-    private fun navigateToSignUp() {
-        val intent = Intent(this, SignUpActivity::class.java)
-        resultLauncher.launch(intent)
-    }
-
-    private fun navigateToMain() {
-        val intent = Intent(this, MainActivity::class.java)
-        intent.putExtra(SIGN_UP_INFORMATION, savedSignUpInformation)
-        startActivity(intent)
-        finish()
     }
 
     private fun setResultSignUpInformation() {
@@ -85,6 +41,29 @@ class SignInActivity : BindingActivity<ActivitySignInBinding>(R.layout.activity_
             }
     }
 
+    private fun setClickEventOnSignUpLabelButton() {
+        binding.btnSignInSignupLabel.setOnClickListener {
+            navigateToSignUp()
+        }
+    }
+
+    private fun navigateToSignUp() {
+        val intent = Intent(this, SignUpActivity::class.java)
+        resultLauncher.launch(intent)
+    }
+
+    private fun setClickEventOnSignInLabelButton() {
+        binding.btnSignInSigninLabel.setOnClickListener {
+            savedSignUpInformation?.let {
+                if (isSignInSuccessful(it)) handleSignInSuccess() else showSignInErrorMessage()
+            }
+        }
+    }
+
+    private fun isSignInSuccessful(user: User): Boolean {
+        return checkIdIdentification(user.id) && checkPwIdentification(user.pw)
+    }
+
     private fun checkIdIdentification(id: String): Boolean {
         return binding.etSignInId.text.toString() == id
     }
@@ -93,5 +72,19 @@ class SignInActivity : BindingActivity<ActivitySignInBinding>(R.layout.activity_
         return binding.etSignInPw.text.toString() == pw
     }
 
+    private fun handleSignInSuccess() {
+        toast(getString(R.string.error_message_valid_sign_in))
+        navigateToMain()
+    }
 
+    private fun navigateToMain() {
+        val intent = Intent(this, MainActivity::class.java)
+        intent.putExtra(SIGN_UP_INFORMATION, savedSignUpInformation)
+        startActivity(intent)
+        finish()
+    }
+
+    private fun showSignInErrorMessage() {
+        snackBar(binding.root) { getString(R.string.error_message_invalid_sign_in) }
+    }
 }
