+++ app/src/main/res/layout/activity_main.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="30dp"
-        tools:context=".features.MainActivity">
+        tools:context=".features.main.MainActivity">
 
         <com.google.android.material.imageview.ShapeableImageView
             android:id="@+id/iv_main_profile"
